+++ src/types/index.ts
@@ -5,3 +5,4 @@ export * from './object-values';
 export * from './parsed-picker-input';
 export * from './picker-options';
 export * from './time-interval';
+export * from './year-picker';